+++ apps/extension/src/ui/domains/Staking/hooks/nomPools/useStakingAPR.ts
@@ -1,28 +1,24 @@
 import { DotNetworkId } from "@talismn/chaindata-provider"
 import { useQuery } from "@tanstack/react-query"
-import { log } from "extension-shared"
 
 import { useScaleApi } from "@ui/hooks/sapi/useScaleApi"
 
 import { getStakingAPR } from "../../helpers"
+import { useBabeNetwork } from "./useBabeNetwork"
 
 export const useStakingAPR = (chainId: DotNetworkId | null | undefined) => {
-  const { data: sapi } = useScaleApi(chainId)
+  const babeNetwork = useBabeNetwork(chainId)
+  const { data: stakingSapi } = useScaleApi(chainId)
+  const { data: babeSapi } = useScaleApi(babeNetwork?.id)
 
   return useQuery({
-    queryKey: ["useStakingAPR", sapi?.id],
-    queryFn: async () => {
-      if (!sapi) return null
+    queryKey: ["useStakingAPR", babeSapi?.id],
+    queryFn: () => {
+      if (!stakingSapi || !babeSapi) return null
 
-      const stop = log.timer(`useStakingAPR(${sapi.chainId})`)
-
-      const apr = await getStakingAPR(sapi)
-
-      stop()
-
-      return apr
+      return getStakingAPR(stakingSapi, babeSapi)
     },
-    enabled: !!sapi,
+    enabled: !!stakingSapi && !!babeSapi,
     refetchInterval: false,
     refetchOnWindowFocus: false,
     refetchOnMount: false,