+++ packages/extension-core/src/domains/app/types.ts
@@ -1,4 +1,4 @@
-import { DotNetworkId, TokenId } from "@talismn/chaindata-provider"
+import { DotNetworkId, EthNetworkId, TokenId } from "@talismn/chaindata-provider"
 
 import { ValidRequests } from "../../libs/requests/types"
 import { Address } from "../../types/base"
@@ -15,6 +15,7 @@ export type RemoteConfigStoreData = {
   swaps: {
     questApi?: string
     lifiApiKey?: string
+    lifiTalismanTokens?: string[]
     simpleswapApiKey?: string
     simpleswapApiKeyDiscounted?: string
     simpleswapDiscountedCurrencies?: string[]
@@ -29,7 +30,7 @@ export type RemoteConfigStoreData = {
     apiUrl: string
   }
   nominationPools: Record<DotNetworkId, number[]>
-  stakingPools: Record<DotNetworkId, (number | string)[]>
+  stakingPools: Record<DotNetworkId | EthNetworkId, (number | string)[]>
   postHogUrl: string
   documentation: {
     unifiedAddressDocsUrl: string
@@ -89,6 +90,7 @@ export type FeatureFlags = Partial<{
   NFTS_V2: boolean
   SEEK_BENEFITS: boolean
   SEEK_TAO_DISCOUNT: boolean
+  SEEK_PRESALE: boolean
 }>
 export type FeatureFlag = keyof FeatureFlags